GH-5810: refactor
