+++ projects/storefrontapp-e2e-cypress/cypress/helpers/wish-list.ts
@@ -1,6 +1,14 @@
-import { cheapProduct } from '../sample-data/checkout-flow';
+import { cheapProduct, user } from '../sample-data/checkout-flow';
 import { login, register } from './auth-forms';
+import * as checkoutAsPersistentUser from './checkout-as-persistent-user';
+import * as checkout from './checkout-flow';
 import { waitForPage } from './checkout-flow';
+import {
+  AddressData,
+  fillPaymentDetails,
+  fillShippingAddress,
+  PaymentDetails,
+} from './checkout-forms';
 import { generateMail, randomString } from './user';
 
 interface TestProduct {
@@ -182,6 +190,99 @@ export function goToProductPage(product: TestProduct) {
     cy.get('.cx-name>.cx-link').click({ force: true });
   });
   cy.wait(`@${productPage}`);
+}
+
+export function checkoutFromWishList(checkoutProducts: TestProduct[]) {
+  goToCartAndCheckout(checkoutProducts);
+  fillAddressForm();
+  checkout.verifyDeliveryMethod();
+  fillPaymentForm();
+  placeOrderWithProducts(checkoutProducts);
+  verifyOrderConfirmationPage(checkoutProducts);
+}
+
+export function checkoutFromCart(checkoutProducts: TestProduct[]) {
+  goToCartAndCheckout(checkoutProducts);
+  checkoutAsPersistentUser.selectShippingAddress();
+  checkoutAsPersistentUser.selectDeliveryMethod();
+  checkoutAsPersistentUser.selectPaymentMethod();
+  placeOrderWithProducts(checkoutProducts);
+  verifyOrderConfirmationPage(checkoutProducts);
+}
+
+function goToCartAndCheckout(checkoutProducts: TestProduct[]) {
+  const cartPage = waitForPage('/cart', 'getCartPage');
+  cy.get('cx-mini-cart').click();
+  cy.wait(`@${cartPage}`);
+
+  for (const product of checkoutProducts) {
+    cy.get('cx-cart-item-list').contains('cx-cart-item', product.code);
+  }
+
+  const shippingAddressPage = waitForPage(
+    '/checkout/shipping-address',
+    'getShippingAddressPage'
+  );
+  cy.getByText(/proceed to checkout/i).click();
+  cy.wait(`@${shippingAddressPage}`);
+}
+
+function fillAddressForm(shippingAddressData: AddressData = user) {
+  cy.get('.cx-checkout-title').should('contain', 'Shipping Address');
+  const deliveryPage = waitForPage(
+    '/checkout/delivery-mode',
+    'getDeliveryPage'
+  );
+  fillShippingAddress(shippingAddressData);
+  cy.wait(`@${deliveryPage}`)
+    .its('status')
+    .should('eq', 200);
+}
+
+function fillPaymentForm(
+  paymentDetailsData: PaymentDetails = user,
+  billingAddress?: AddressData
+) {
+  cy.get('.cx-checkout-title').should('contain', 'Payment');
+  const reviewPage = waitForPage('/checkout/review-order', 'getReviewPage');
+  fillPaymentDetails(paymentDetailsData, billingAddress);
+  cy.wait(`@${reviewPage}`)
+    .its('status')
+    .should('eq', 200);
+}
+
+function placeOrderWithProducts(checkoutProducts: TestProduct[]) {
+  cy.get('.cx-review-title').should('contain', 'Review');
+
+  for (const product of checkoutProducts) {
+    cy.get('cx-cart-item-list').contains('cx-cart-item', product.code);
+  }
+
+  cy.getByText('Terms & Conditions')
+    .should('have.attr', 'target', '_blank')
+    .should(
+      'have.attr',
+      'href',
+      '/electronics-spa/en/USD/terms-and-conditions'
+    );
+  cy.get('.form-check-input').check();
+  const orderConfirmationPage = waitForPage(
+    '/order-confirmation',
+    'getOrderConfirmationPage'
+  );
+  cy.get('cx-place-order button.btn-primary').click();
+  cy.wait(`@${orderConfirmationPage}`)
+    .its('status')
+    .should('eq', 200);
+}
+
+function verifyOrderConfirmationPage(checkoutProducts: TestProduct[]) {
+  cy.get('.cx-page-title').should('contain', 'Confirmation of Order');
+  cy.get('h2').should('contain', 'Thank you for your order!');
+
+  for (const product of checkoutProducts) {
+    cy.get('cx-cart-item-list').contains('cx-cart-item', product.code);
+  }
 }
 
 function getCartItem(name: string) {
